// src/types/api.types.ts
export type Article = {
  id: number;
  title: string;
  slug: string;
  content: string;
  is_published: boolean;
  view_count: number;
  hashtags: string[];
  admin_id: number;
  description:string;
  // created_at: string;
  // updated_at: string;
  image: string
};